requirements.txt
beautifulsoup4==4.6.0
certifi==2019.6.16
chardet==3.0.4
Django==2.1.11
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.2
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
six==1.12.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.6.1
wagtailfontawesome==1.1.4
webencodings==0.5.1
Willow==1.1
